day3/Discord_MCP_Server/requirements.txt
# FastMCP - Main framework for MCP server
fastmcp>=2.0.0
# The main framework that makes our MCP server work

# Discord integration
discord.py>=2.3.0
#  Connects to Discord API

# Environment variables
python-dotenv>=1.0.0

# Async HTTP requests
aiohttp>=3.8.0
# Handles async HTTP requests

# Logging
loguru>=0.7.0

# JSON handling
orjson>=3.9.0
#  Fast JSON processing

# Date/time handling
python-dateutil>=2.8.0

# Testing (for later)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.0.0 